requirements.txt
click==7.1.2
    # via typer
pyopenal==0.7.11a1
    # via -r requirements.in
typer==0.3.2
    # via -r requirements.in
